microservices/lambda/orchestation_service/requirements.txt
fastapi
uvicorn
boto3
requests
python-dotenv
mangum
